rename bench_merge.py to bench.py
