+++ bench_toolbox/configs/parsers/__init__.py
@@ -1,6 +1,6 @@
 # coding: UTF-8
 
 from .base import BaseParser, LocalReadParser
-from .bench_merger import BenchParser
+from .bench import BenchParser
 from .perf import PerfParser
 from .rabbit_mq import RabbitMQParser
